+++ src/common/AppUpdateBanner.tsx
@@ -1,7 +1,6 @@
 import { faSyncAlt as reloadIcon } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { useToggle } from '@shlinkio/shlink-frontend-kit';
-import { Button, Card, CloseButton } from '@shlinkio/shlink-frontend-kit/tailwind';
+import { Button, Card, CloseButton,useToggle  } from '@shlinkio/shlink-frontend-kit';
 import { clsx } from 'clsx';
 import type { FC } from 'react';
 import { useCallback } from 'react';
@@ -13,7 +12,7 @@ interface AppUpdateBannerProps {
 }
 
 export const AppUpdateBanner: FC<AppUpdateBannerProps> = ({ isOpen, onClose, forceUpdate }) => {
-  const [isUpdating,, setUpdating] = useToggle();
+  const { flag: isUpdating, setToTrue: setUpdating } = useToggle();
   const update = useCallback(() => {
     setUpdating();
     forceUpdate();
@@ -27,15 +26,15 @@ export const AppUpdateBanner: FC<AppUpdateBannerProps> = ({ isOpen, onClose, for
     <Card
       role="alert"
       className={clsx(
-        'tw:w-[700px] tw:max-w-[calc(100%-30px)]',
-        'tw:fixed tw:top-[35px] tw:left-[50%] tw:translate-x-[-50%] tw:z-[1040]',
+        'w-[700px] max-w-[calc(100%-30px)]',
+        'fixed top-[35px] left-[50%] translate-x-[-50%] z-[1040]',
       )}
     >
-      <Card.Header className="tw:flex tw:items-center tw:justify-between">
+      <Card.Header className="flex items-center justify-between">
         <h5>This app has just been updated!</h5>
         <CloseButton onClick={onClose} />
       </Card.Header>
-      <Card.Body className="tw:flex tw:gap-4 tw:items-center tw:justify-between tw:max-md:flex-col">
+      <Card.Body className="flex gap-4 items-center justify-between max-md:flex-col">
         Restart it to enjoy the new features.
         <Button disabled={isUpdating} variant="secondary" solid onClick={update}>
           {!isUpdating && <>Restart now <FontAwesomeIcon icon={reloadIcon} /></>}