+++ src/common/ErrorLayout.tsx
@@ -1,4 +1,4 @@
-import { SimpleCard } from '@shlinkio/shlink-frontend-kit/tailwind';
+import { SimpleCard } from '@shlinkio/shlink-frontend-kit';
 import type { FC, PropsWithChildren } from 'react';
 
 export type ErrorLayoutProps = PropsWithChildren<{
@@ -6,8 +6,8 @@ export type ErrorLayoutProps = PropsWithChildren<{
 }>;
 
 export const ErrorLayout: FC<ErrorLayoutProps> = ({ children, title }) => (
-  <div className="tw:pt-4">
-    <SimpleCard className="tw:p-4 tw:w-full tw:lg:w-[65%] tw:m-auto">
+  <div className="pt-4">
+    <SimpleCard className="p-4 w-full lg:w-[65%] m-auto">
       <h2>{title}</h2>
       {children}
     </SimpleCard>