+++ vite.config.ts
@@ -21,9 +21,11 @@ export default defineConfig({
     manifestFilename: 'manifest.json',
     manifest,
   })],
+
   build: {
     outDir: 'build',
   },
+
   server: {
     port: 3000,
     watch: {
@@ -31,6 +33,7 @@ export default defineConfig({
       ignored: ['**/.idea/**', '**/.git/**', '**/build/**', '**/coverage/**', '**/test/**'],
     },
   },
+
   base: !homepage ? undefined : homepage, // Not using just homepage because empty string should be discarded
 
   // Vitest config